+++ src/test/java/io/github/classgraph/issues/issue246/Issue246Test.java
@@ -49,7 +49,7 @@ public class Issue246Test {
                             .getMethodInfo() //
                             .getSingleMethod("dummyMethodWithParameter") //
                             .getParameterInfo()[0] //
-                            .getAnnotationInfo().size());
+                            .getAllAnnotationInfo().size());
         }
     }